requirements.txt
uvicorn==0.11.8
fastapi==0.83.0
autopep8==2.0.2
requests==2.28.1
pydantic==1.10.7
pymongo==3.11.0
PyJWT==2.7.0
dnspython==2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
pinecone-client==2.2.2
sentence-transformers==2.2.2
langchain
openai==0.27.8
boto3==1.26.136
botocore==1.29.136
Pillow==9.5.0
secure-smtplib==0.1.1
mailtrap==2.0.1
sacremoses==0.0.53